every features done .

+++ admn/views.py
@@ -1,6 +1,8 @@
+
 from django.contrib import auth, messages
 from django.contrib.auth.models import User
 from django.shortcuts import render, redirect
+from django.template.defaultfilters import lower
 from django.views.decorators.cache import never_cache
 
 from admn.forms import newuser
@@ -14,7 +16,8 @@ def admin(request):
             Cpassw = request.COOKIES.get('admin-password')
             Sname = request.session['admin-ID']
             Spassw = request.session['admin-password']
-            if Cname == Sname and Cpassw == Spassw:
+            Usr = auth.authenticate(username=Cname, password=Cpassw)
+            if Cname == Sname and Cpassw == Spassw and Usr:
                 return redirect('admin/home')
             else:
                 return render(request, 'admin-form.html', {'ttl': "Admin-Form"})
@@ -27,21 +30,21 @@ def admin(request):
 def login(request):
     if request.method == 'POST':
         if request.POST['adminid'] and request.POST['AdminPassword'] :
-            uname = request.POST['adminid']
+            adname = request.POST['adminid']
             password = request.POST['AdminPassword']
-            usr = auth.authenticate(username=uname,password=password)
-            istaf = User.objects.get(username=uname)
+            usr = auth.authenticate(username=adname,password=password)
+            istaf = User.objects.get(username=adname)
             if usr and istaf.is_staff==True:
                 auth.login(request,usr)
                 res = redirect('admin/home')
-                res.set_cookie('admin-ID', uname)
+                res.set_cookie('admin-ID', adname)
                 res.set_cookie('admin-password', password)
-                request.session['admin-ID'] = uname
+                request.session['admin-ID'] = adname
                 request.session['admin-password'] = password
                 return res
 
             else:
-                messages.error(request,'Invalid credentials or Youare not an admin')
+                messages.error(request,'Invalid credentials or You are not an admin')
                 return redirect('/admin')
         else:
             messages.error(request, 'you cannot submit without credentials filled !')
@@ -59,7 +62,14 @@ def adminhome(request):
         usr = auth.authenticate(username=Cname, password=Cpassw,is_staff="true")
         if usr:
             if Cname == Sname and Cpassw == Spassw:
-                users = User.objects.filter()
+                if request.method=='GET':
+                    s= lower( request.GET['search'])
+                    if s == "" or s== "search by user name":
+                        users = User.objects.filter()
+                    else:
+                        users = User.objects.filter(username=s)
+                else:
+                    users = User.objects.filter()
                 form = newuser()
                 return render(request,'admin-home.html',{"name":Cname,'users':users,'form':form})
 
@@ -83,9 +93,10 @@ def show(request,Usrid):
     Cname = request.COOKIES.get('admin-ID')
     users = User.objects.get(id=Usrid)
     Username = users.username
+    form = newuser()
     td = todo.objects.filter(username=Username, task_status=False)
     dn = todo.objects.filter(username=Username, task_status=True)
-    return render(request,'user-data.html',{'td':td, 'dn':dn,'username':Username,'name':Cname})
+    return render(request,'user-data.html',{'td':td, 'dn':dn,'username':Username,'name':Cname,'form':form})
 
 @never_cache
 def delete(request,taskid):
@@ -113,19 +124,16 @@ def done(request, taskid):
 
 def adduser(request):
     if request.method == 'POST':
-        if request.POST['username'] and request.POST['password']:
+        if not request.POST['username'] and not request.POST['password']:
+            messages.error(request, 'you cannot submit without credentials filled !')
+            return redirect('/admin')
+        else:
             uname = request.POST['username']
             passw = request.POST['password']
-            if request.POST['is_staff']:
-                staf = True
-            else:
-                staf = False
+            staf = request.POST['is_staff']
             usr = User.objects.create_user(username=uname, password=passw,is_staff=staf)
             usr.save()
             print("user sign inned successfully")
-            return redirect('/admin')
-        else:
-            messages.error(request, 'you cannot submit without credentials filled !')
             return redirect('/admin')
     else:
         return redirect('/admin')
@@ -133,5 +141,24 @@ def adduser(request):
 
 def deleteuser(request,userid):
     if request.method == 'GET':
-        User.objects.get(id=userid).delete()
+        usr = User.objects.get(id=userid)
+        if usr.is_superuser == False :
+            User.objects.get(id=userid).delete()
+            return redirect('/admin')
+        else:
+            messages.error(request, 'You cannot delete a Super User !')
+            return redirect('/admin')
+
+
+def update(request):
+    if request.method == 'POST':
+        oldusername = request.POST['oldusername']
+        newusername = request.POST['username']
+        password = request.POST['password']
+        staf = request.POST['is_staff']
+        usr=User.objects.get(username=oldusername)
+        usr.delete()
+        usr = User.objects.create_user(username=newusername, password=password, is_staff=staf)
+        usr.save()
         return redirect('/admin')
+
